+++ apps/db-service/package.json
@@ -20,8 +20,8 @@
     "@nestjs/common": "^10.0.0",
     "@nestjs/core": "^10.0.0",
     "@nestjs/platform-express": "^10.0.0",
-    "@prisma/client": "^6.15.0",
-    "@trading-journal/shared": "*",
+    "@prisma/client": "^6.13.0",
+    "@trading-journal/types": "*",
     "reflect-metadata": "^0.2.0",
     "rxjs": "^7.8.1"
   },
