+++ apps/db-service/prisma/schema.prisma
@@ -5,7 +5,7 @@ datasource db {
 
 generator client {
     provider = "prisma-client-js"
-    output   = "./generated/prisma"
+    output   = "../node_modules/@prisma/client"
 }
 
 model Trade {
