+++ apps/db-service/src/modules/trades/trades.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable } from "@nestjs/common";
 
-import { Trade } from "@trading-journal/shared";
+import { Trade } from "@trading-journal/types";
 import { DatabaseRepository } from "../database/database.repository.js";
 
 @Injectable()
