+++ apps/frontend/src/components/Tabs/Tabs.tsx
@@ -1,5 +1,5 @@
 import { Box } from "@mui/material";
-import type { Trade } from "@trading-journal/shared";
+import type { Trade } from "@trading-journal/types";
 import { useState } from "react";
 import EmptyState from "../EmptyState";
 import TradeList from "../TradeList/TradeList";
