+++ apps/frontend/src/components/TradeForm/TradeFormEntryContainer/TradeFormEntryContainer.tsx
@@ -1,5 +1,5 @@
 import { Box, Grid } from "@mui/material";
-import { type NewTradeFields, AddTradeEntryLabels } from "@trading-journal/shared";
+import { type NewTradeFields, AddTradeEntryLabels } from "@trading-journal/types";
 import { type Control, type FieldErrors } from "react-hook-form";
 import ControlledDatePicker from "../../ControlledComponents/ControlledDatePicker";
 import ControlledTextField from "../../ControlledComponents/ControlledTextField";
