+++ apps/frontend/src/components/TradeList/TradeTotalsDisplay/TradeTotalsDisplay.tsx
@@ -1,5 +1,5 @@
 import { Box, Typography } from "@mui/material";
-import type { Trade } from "@trading-journal/shared";
+import type { Trade } from "@trading-journal/types";
 
 interface TradeTotalDisplayProps {
     pnl: Trade["pnl"],
