+++ apps/frontend/tsconfig.json
@@ -17,7 +17,7 @@
     "baseUrl": "./src",
     "paths": {
       "src/*": ["./*"],
-      "@trading-journal/shared/*": ["./src/*"]
+      "@trading-journal/types/*": ["./src/*"]
     }
   }
 }
