+++ package.json
@@ -17,7 +17,7 @@
   "license": "ISC",
   "packageManager": "npm@10.8.1",
   "dependencies": {
-    "@trading-journal/shared": "*"
+    "@trading-journal/types": "*"
   },
   "devDependencies": {
     "tsconfig-paths": "^4.2.0",
