+++ packages/types/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@trading-journal/shared",
+    "name": "@trading-journal/types",
     "version": "0.0.1",
     "main": "src/index.ts",
     "scripts": {
