+++ packages/types/src/index.ts
@@ -1,3 +1,3 @@
 export * from './enums/index';
-export * from './types/index';
+export * from './trades/index';
